% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_admits.R
\name{extract_admits}
\alias{extract_admits}
\title{Extract admits by nursing unit}
\usage{
extract_admits(startdate_id, enddate_id, site = "Lions Gate Hospital",
  n_units = c("LGH 4E"), denodo_vw = vw_adtc)
}
\arguments{
\item{startdate_id}{e.g. "20190101"}

\item{enddate_id}{e.g. "20190101"}

\item{site}{e.g. "Lions Gate Hospital"}

\item{n_units}{e.g. "LGH 4E"}

\item{denodo_vw}{vw_adtc}
}
\value{
a dataframe
}
\description{
Extract admits by nursing unit
}
