% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dates.R
\name{fill_dates}
\alias{fill_dates}
\title{Fill in missing dates in a dataframe with intermittent dates}
\usage{
fill_dates(df, date_col, start_date, end_date)
}
\arguments{
\item{df}{a dataframe where at least 1 column has dates}

\item{date_col}{name of the column that has dates}

\item{start_date}{start date, e.g. "2019-01-01"}

\item{end_date}{end date, e.g. "2019-02-01"}
}
\value{
the dataframe you started with, expanded to include all missing dates as rows
}
\description{
Fill in missing dates in a dataframe with intermittent dates
}
\examples{
df <- data.frame(dates = c("2019-01-01", "2019-02-01"), numbers = c(3,4))
df \%>\% fill_dates(dates, "2019-01-01", "2019-02-08")
}
