% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_denodo.R
\name{setup_denodo}
\alias{setup_denodo}
\title{Set up ODBC connection to denodo}
\usage{
setup_denodo(dsn_name = "cnx_denodo_spappcsta001")
}
\arguments{
\item{dsn_name}{A string specifying connection name. Default value is "cnx_denodo_spappcsta001", because that's the name I use.}
}
\value{
Returns connections to databases
}
\description{
Set up ODBC connection to denodo
}
\examples{
setup_denodo()
# should return 3 views: vw_adtc, vw_census, vw_eddata
}
