% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sql_server.R
\name{setup_sql_server}
\alias{setup_sql_server}
\title{Set up ODBC connection to SQL Server}
\usage{
setup_sql_server(dsn_name = "cnx_SPDBSCSTA001")
}
\arguments{
\item{dsn_name}{A string specifying connection name. Default value is "cnx_SPDBSCSTA001", because that's the name I use.}
}
\value{
Returns connections to databases
}
\description{
Set up ODBC connection to SQL Server
}
\examples{
setup_sql_server()
# Should return views with names like vw_ed_mart, vw_adtc_ad_view, etc.
}
